added downstreamKeyer function
