+++ src/feedback.ts
@@ -8,6 +8,8 @@ import { StreamingFeedbacks } from './functions/streaming'
 import { LiveFeedbacks } from './functions/live'
 import { RecordFeedbacks } from './functions/record'
 import { StillGeneratorFeedbacks } from './functions/stillGenerator'
+import { DownstreamKeyerFeedbacks } from './functions/downstreamKeyer'
+
 import {
 	UpStreamKeyTypeChoices,
 	SettingsInputWindowLayoutChoices,
@@ -31,31 +33,6 @@ export const MacroFeedbackType = {
 
 export function feedbacks(instance: GoStreamInstance): CompanionFeedbackDefinitions {
 	return {
-		[feedbackId.DskOnAir]: {
-			type: 'boolean',
-			name: 'Next Transition:DSK OnAir',
-			description: 'Set the special effect Transition DSK OnAir',
-			options: [
-				{
-					type: 'dropdown',
-					label: 'DSK OnAir',
-					id: 'DSKOnAir',
-					choices: SwitchChoices,
-					default: 1,
-				},
-			],
-			defaultStyle: {
-				color: combineRgb(0, 0, 0),
-				bgcolor: combineRgb(255, 255, 0),
-			},
-			callback: (feedback) => {
-				if (instance.states.TKeyeState.DSKOnAir && feedback.options.DSKOnAir === 1) {
-					return true
-				} else {
-					return false
-				}
-			},
-		},
 		[feedbackId.TransitionSource]: {
 			type: 'boolean',
 			name: 'USK: Tied',
@@ -161,43 +138,6 @@ export function feedbacks(instance: GoStreamInstance): CompanionFeedbackDefiniti
 			callback: (feedback) => {
 				const typeId = Number(feedback.options.USKKeyType)
 				return instance.states.upStreamKeyState.ArrayKeySourceFill[typeId] === Number(feedback.options.USKSourceFill)
-			},
-		},
-		[feedbackId.DskSourceFill]: {
-			type: 'boolean',
-			name: 'DSK:DSK source fill or key fill',
-			description: 'Set the special effect DSK source fill or key fill',
-			options: [
-				{
-					type: 'dropdown',
-					label: 'Type:',
-					id: 'TypeID',
-					choices: [
-						{ id: 0, label: 'Key Fill' },
-						{ id: 1, label: 'Source Fill' },
-					],
-					default: 0,
-				},
-				{
-					type: 'dropdown',
-					label: 'Fill:',
-					id: 'DSKFill',
-					choices: getChoices(ActionType.KeyPatternSourceKey),
-					default: 0,
-				},
-			],
-			defaultStyle: {
-				color: combineRgb(0, 0, 0),
-				bgcolor: combineRgb(255, 255, 0),
-			},
-			callback: (feedback) => {
-				const typeId = Number(feedback.options.TypeID)
-				const dsk_source = Number(feedback.options.DSKFill)
-				if (typeId === 0) {
-					return instance.states.DSKState.DSKSourceKeyFill.id === dsk_source
-				} else {
-					return instance.states.DSKState.DSKSourceFill.id === dsk_source
-				}
 			},
 		},
 		[feedbackId.InputWindowLayout]: {
@@ -653,5 +593,6 @@ export function feedbacks(instance: GoStreamInstance): CompanionFeedbackDefiniti
 		...MixEffectFeedbacks.create(instance),
 		...StreamingFeedbacks.create(instance),
 		...StillGeneratorFeedbacks.create(instance),
+		...DownstreamKeyerFeedbacks.create(instance),
 	}
 }
