+++ src/functions/colorBack/actionId.ts
@@ -1,0 +1,5 @@
+export enum ActionId {
+	ColorHue = 'colorHue',
+	ColorSaturation = 'colorSaturation',
+	ColorBrightness = 'colorBrightness',
+}
