+++ src/functions/colorBack/index.ts
@@ -1,0 +1,1 @@
+export * as ColorBackActions from './actions'
