+++ src/functions/downstreamKeyer/actionId.ts
@@ -1,0 +1,16 @@
+export enum ActionId {
+	DskOnAir = 'dskOnAir',
+	DskSourceFillKey = 'dskSourceFillKey',
+	DskSourceFill = 'dskSourceFill',
+	DskSourceKey = 'dskSourceKey',
+	DskMaskEnable = 'dskMaskEnable',
+	DskMaskHStart = 'dskMaskHStart',
+	DskMaskVStart = 'dskMaskVStart',
+	DskMaskHEnd = 'dskMaskHEnd',
+	DskMaskVEnd = 'dskMaskVEnd',
+	DskControlShapedKey = 'dskControlShapedKey',
+	DskControlClip = 'dskControlClip',
+	DskControlGain = 'dskControlGain',
+	DskControlInvert = 'dskControlInvert',
+	DskRate = 'dskRate',
+}
