+++ src/functions/downstreamKeyer/feedbackId.ts
@@ -1,0 +1,4 @@
+export enum FeedbackId {
+    DskOnAir = 'dskOnAir',
+    DskSourceFill = 'dskSourceFill',
+}
